Add delete

+++ app/src/main/java/com/ridecam/fs/FSUtils.java
@@ -56,6 +56,7 @@ public class FSUtils {
     }
 
     public static void deleteFile(File file) {
+        if (!file.exists()) return;
         file.delete();
     }
